Split out core functions
